migrate to npm plugin usage

+++ backend/src/database.rs
@@ -18,11 +18,15 @@ impl SQLiteDatabase {
             [],
         );
 
-        Ok(SQLiteDatabase {
+        let self_inst = SQLiteDatabase {
             connection,
             last_update: 0,
             min_time_between_cleans: min_time_between_cleans,
-        })
+        };
+
+        let _ = self_inst.clear_database();
+
+        Ok(self_inst)
     }
 
     pub fn get_values(
@@ -111,6 +115,10 @@ impl SQLiteDatabase {
     }
 
     fn clean_database_time(&self, min_time_since_last_update: u32) -> Result<(), rusqlite::Error> {
+        if self.last_update < min_time_since_last_update {
+            return Ok(());
+        }
+
         self.connection.execute(
             "DELETE FROM data WHERE timestamp <= ?",
             [(self.last_update - min_time_since_last_update).to_string()],
